+++ spec/fixtures/wait-2-seconds.js
@@ -1,0 +1,3 @@
+setTimeout(function() {
+  console.log('2 seconds have passed')
+}, 2000)