+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+
+* Add `uniqueKey` support to exec options
+
 ## 9.0.1
 
 *  Make `exec` and `execNode` behave the way they were supposed and documented to be by restoring their 10 seconds default timeout